+++ packages/react-native-sdk/__tests__/mocks/call.ts
@@ -8,6 +8,7 @@ import mockParticipant from './participant';
 export const mockCall = (
   client: StreamVideoClient,
   participants?: StreamVideoParticipant[],
+  ownCapabilities?: OwnCapability[],
 ) => {
   const call = client?.call('default', 'test-123');
   const _participants = participants || [mockParticipant()];
@@ -17,5 +18,10 @@ export const mockCall = (
     OwnCapability.SEND_AUDIO,
     OwnCapability.SEND_VIDEO,
   ]);
+  call.permissionsContext.setPermissions(ownCapabilities ?? []);
+  call.state.setOwnCapabilities(ownCapabilities ?? []);
+
+  // To handle the hang up call we need to mock the `leave` function
+  call.leave = jest.fn();
   return call;
 };