+++ sample-apps/react/messenger-clone/src/utils/constants.ts
@@ -1,0 +1,2 @@
+export const SESSION_STORAGE_KEY = '@stream-messenger-clone';
+export const SESSION_STORAGE_USER_KEY = `${SESSION_STORAGE_KEY}/user`;