+++ sample-apps/react/zoom-clone/src/utils/constants.ts
@@ -1,2 +1,4 @@
-export const SESSION_STORAGE_KEY = 'zc:uid';
+export const SESSION_STORAGE_KEY = '@stream-audio-room';
+export const SESSION_STORAGE_USER_KEY = `${SESSION_STORAGE_KEY}/user`;
 export const DEFAULT_CHANNEL_TYPE = 'videocall';
+export const DEFAULT_CALL_TYPE = 'default';